Move public position function
